add 15 minute ttl to gauges
after setting a gauge, it's value will now remain set in the registry for 15
minutes.

this brings the python client in-line with existing functionality in the java
reference client.

previously, this ttl behavior was provided transitively through the aggregators,
but the ttls were reduced on that side to improve the behavior associated with
averaging gauges when instances are terminated.