post_json should retry when an HTTP error occurs
We recently observed a case where the ELB in front of the aggregator
service was overloaded with traffic and it was returning 503 errors
periodically. This change will retry posts up to three times on HTTPError
exceptions, with a three second delay between retries.

We are opting not to implement a `logentry` class, which is responsible for
creating common IPC metrics for these post calls, due to some differences
between the spectator-py and spectator-js implementations that will need to
be hashed out.

The URLError exception block was nearly identical to the generic exception
block; they have been folded together. URLErrors are not likely to succeed
upon retries, so they do not.

A handful of methods have been annotated as static methods, since they were
not making use of class instance data.

Decode calls were added to `_read_response` and `_read_error`, so that the
log messages do not show bytes literal notation.