add python 3.5
We wanted to move up to Python 3.6 as a minimum version, so that we could
start taking advantage of f-strings, but we also want this library to work
with the minimum number of dependencies outside of the system Python 3
provided by our chosen Linux distributions.

The oldest system Python in our base OS collection is 3.5, which won't be
retired until sometime in 2021. The next oldest at that point will be 3.6.