add SPECTATOR_OUTPUT_LOCATION environment variable config
We have a case where we want to disable sending messages to the udp socket
when running integration tests on CI servers which also happen to be running
SpectatorD. The applications under test use the GlobalRegistry liberally.

The SPECTATOR_OUTPUT_LOCATION environment variable may be set to "none" to
disable publishing messages, or "memory" if there is a need to verify the
messages.