fix protected access
In order to be more consistent about how the MemoryWriter is accessed
during testing, convenience methods have been added to the Registry
and the SidecarMeter classes which return the internal SidecarWriter.

These methods were annotated with a Union type, because that appears
to be the cleanest way to get PyCharm to acknowledge that the return
value can be one of the four SidecarWriter subclasses.

With this change, there are no longer any protected access warnings
for this project.