remove release status badge
GitHub Actions status badges do not currently support the ability to track
dynamic tag versions on a branch. In order to track releases, we would have
to update the Release badge for each tagged version, like so:

```
[![Release](https://github.com/Netflix/spectator-py/actions/workflows/release.yml/badge.svg?branch=v0.2.6)](https://github.com/Netflix/spectator-py/actions/workflows/release.yml)
```

This is a bit too much overhead for releases, so we'll drop this badge for
now. According to GitHub support, there is an internal feature request to
support this sort of tag tracking. If and when they release such a feature,
we can add it back.

In the meantime, tracking the Snapshot status, which occurs on merge to the
`main` branch, should be sufficient to reflect the project build status.