improve asyncio support in UdpWriter
The previous method of opening a socket during the UdpWriter init and re-using
it for sending packets was not universally usable with asyncio, because the
initial call was blocking, even though the send calls were non-blocking. This
meant that the initial import of the GlobalRegistry from spectator-py needed
to be handled outside of async code, to avoid BlockingIO errors, but otherwise,
metrics data would be sent as expected. This implementation was originally
selected, because it was thought that it performed slightly better than the
alternative.

This commit changes the implementation of the UdpWriter, so that a new socket
is opened within a context manager for every message that is sent to the
SpectatorD sidecar. The context manager ensures that the socket is closed
after the message is sent, so that we do not run out of file descriptors for
the process. These calls are all non-blocking, which means that the library
can now be used generally within asyncio applications, without needing to
consider where the initial import occurs. Re-testing with this method shows
that the performance is close to the original method.

The README is updated with advice on asyncio support.