filter invalid tags (#51)
When using this library and setting invalid tags, it was raising errors
on the creation of the SpectatorD protocol string. This makes the library
a little less robust than we would like - it is best for reliability if
this library stays out of the way, in the event that it receives invalid
data, so that we minimize the potential blast radius for services that
use this library.

This change will intercept, modify and log any invalid meters that are
created. Log messages will be written to the `spectator.MeterId` logger at
the warning level at MeterId creation. All not-a-str values will be replaced
with empty strings, to ensure that a valid SpectatorD line protocol message
can be generated without raising an exception. The resulting message will fail
Atlas publishing validation, which will produce a SpectatorD error log and an
increment in the spectatord.measurements metric.