enable ipv6-only support
The current implementation works for dual-stacked systems, but if we start
seeing more ipv6-only systems, then we need to be able to pick the socket
family correctly, given a configuration override.

This change does a check on the udp address to determine the family and
uses that value for future connections. The readme has been updated with
an explanation of how the current communication path works and how to set
and override for ipv6-only systems.