cleanup api for 1.0.0 release
The last release of this library that was a thick client, which published
directly to Atlas instead of spectatord, was v0.1.18 in Apr 2022. The last
release of this library as a thin client before the 1.0 release is v0.2.10
in Jan 2023.

As we have made progress towards our vision of making all the metrics library
clients thin, so that they report directly to spectatord, where all of the
complex logic related to Atlas publishing resides, we have more examples of
what these libraries should look like.

We recently completed the spectator-go thin client library migration, and it
turned out well. Taking some inspiration from that rewrite, and with a desire
to remove the last vestigial bits of the thick client from this library, we
want to cleanup the API for a 1.0 release.

With this PR, the overall design of library is better, less complicated, and
easier to read. One of the goals of this update was to minimize the amount of
change that needs to be absorbed between the v0.2.10 release and a v1.0.0
release, and that has been largely achieved. The `Registry` is the primary
user interface to the library, and those method signatures are unchanged, and
a few new ones have been added, to allow working with Ids directly.

Thus, the migration advice for moving from 0.1.X to 0.2.X should hold for the
1.0.0 release, but we will release it as a candidate first, to validate that
assertion.

* https://netflix.github.io/atlas-docs/spectator/lang/py/migrations/

This change makes the library more similar to spectator-go, and between these
two libraries, a solid pattern exists for moving the spectator-js library to a
thin client implementation.

* https://github.com/Netflix/spectator-go
* https://github.com/Netflix/spectator-js