implement feedback (#66)
* Rename the `Id` class back to `MeterId`, because `Id` is overloaded
and Python has a builtin `id`. This will help keep the uses of this
class a distinct concept.
* Remove the `Meter` logger, because it was unused. All of the important
logging happens in other places, such as the `MeterId` or the `Writer`.
* Rename all the loggers to `__name__`, because the Python logging
module automatically keeps track of the hierarchy. For example, the
`MeterId` logger name becomes `spectator.meter.meter_id`.
* Clarify the `extra_common_tags` concept, with parameter name changes,
and docstrings.
* Restore the `StopWatch` class, which is provided as a convenience to
developers, so that they do not need to recreate this everywhere that
they want to use a Context Manager for timing functions. In this version,
it is separated from the `PercentileTimer` and `Timer` meters, but it will
support both types.
* Restore the `Clock`, `ManualClock`, and `SystemClock` classes, for the
purpose of providing deterministic testing for the `StopWatch` class. These
classes remain separated from the `Meter` classes, which no longer need it.