fix socket communications
For the `UdpWriter`, only open the socket once, during initialization. While
this has not been an issue in practice, and performs comparably, it leads to
port churn on the host that uses this. A single open socket will lead to a
single port being used for communication.

The `FileWriter` is not sufficient for communicating with Unix Domain Sockets,
and we need to add a `UnixWriter` to handle this correctly. Verified in tests.