lazily open sockets
The legacy `GlobalRegistry` will open sockets upon import (`UdpWriter` by
default). For applications that spawn a lot of threads, we want to keep
this behavior under control, so lazily instantiate the sockets for the
`UnixWriter` and the `UdpWriter`. We implement a `threading.Lock` when
instantiating the sockets, to help avoid races between threads.