fixes for mypy type checking (#87)
Since we want to add the `py.typed` marker to this library, we need to check
and fix the types.

This change will enable the following pull request:

https://github.com/Netflix/spectator-py/pull/86

Due to how we are using TypeAliases, we need a Python >= 3.10. Given the status
of Python versions, with 3.8 and 3.9 being end-of-life, we have updated the
supported versions, and set the minimum version to 3.10.

https://devguide.python.org/versions/